:art: Add detailed package overview in readme.md
A comprehensive overview of the project package has been added to the readme.md file. This includes steps to understanding the function and structure of the package, as well as a navigation guide to the example directory. A concise description of what learners may gain from exploring the examples folder and the tutorials therein is also provided. Redundant separations between step points under .